Custom hook fetch data ans used it
